requirements.txt
Flask==0.10.1
Flask-Admin==1.0.9
Flask-Login==0.2.11
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.7.9
WTForms==2.0.2
Werkzeug==0.10.1
argparse==1.2.1
decorator==3.4.0
itsdangerous==0.24
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
